+++ services/weather/internal/adapter/weather/openweathermap/models.go
@@ -3,7 +3,7 @@ package openweathermap
 type Response struct {
 	Main struct {
 		Temp     float64 `json:"temp"`
-		Humidity int     `json:"humidity"`
+		Humidity float64 `json:"humidity"`
 	} `json:"main"`
 	Weather []struct {
 		Description string `json:"description"`
