+++ services/weather/internal/core/domain/weather.go
@@ -5,7 +5,7 @@ import "time"
 type Weather struct {
 	City        string    `json:"city"`
 	Temperature float64   `json:"temperature"`
-	Humidity    int       `json:"humidity"`
+	Humidity    float64   `json:"humidity"`
 	Description string    `json:"description"`
 	WindSpeed   float64   `json:"wind_speed"`
 	Timestamp   time.Time `json:"timestamp"`
